chore: move files to remote dir
