+++ README.md
@@ -27,7 +27,7 @@ https://github.com/mattzcarey/code-review-gpt/assets/77928207/92029baf-f691-465f
 
 ## Prerequisites
 
-- Node.js
+- Node 18+
 - Git
 - Github or Gitlab CLI (optional for configure tool)
