+++ src/common/remote/github/isEligibleForReview.ts
@@ -1,12 +1,12 @@
 import { extname } from "path";
-import { excludedKeywords, supportedFiles } from "../../review/constants";
+import { excludedKeywords, supportedFiles } from "../../../review/constants";
 
 // see: https://docs.github.com/en/rest/pulls/pulls?apiVersion=2022-11-28#list-pull-requests-files
 const excludedFileChangeTypes = ["removed", "unchanged"];
 
 const isSupportedPullRequestFileChange = (status: string): boolean => {
   return !excludedFileChangeTypes.includes(status);
-}
+};
 
 const isSupportedFileType = (path: string): boolean => {
   const ext = extname(path);
@@ -14,8 +14,8 @@ const isSupportedFileType = (path: string): boolean => {
     supportedFiles.has(ext) &&
     ![...excludedKeywords].some((keyword) => path.includes(keyword))
   );
-}
+};
 
 export const isEligibleForReview = (path: string, status: string): boolean => {
   return isSupportedPullRequestFileChange(status) && isSupportedFileType(path);
-}
+};
